+++ app/src/main/java/com/brandoncano/resistancecalculator/navigation/Navigation.kt
@@ -10,6 +10,12 @@ import androidx.navigation.compose.NavHost
 import androidx.navigation.compose.composable
 import androidx.navigation.compose.rememberNavController
 import com.brandoncano.resistancecalculator.constants.Links
+import com.brandoncano.resistancecalculator.navigation.calculators.colorToValueScreen
+import com.brandoncano.resistancecalculator.navigation.calculators.smdScreen
+import com.brandoncano.resistancecalculator.navigation.calculators.valueToColorScreen
+import com.brandoncano.resistancecalculator.navigation.learn.learnColorCodes
+import com.brandoncano.resistancecalculator.navigation.learn.learnPreferredValues
+import com.brandoncano.resistancecalculator.navigation.learn.learnSmdCodes
 import com.brandoncano.sharedcomponents.data.Apps
 import com.brandoncano.sharedcomponents.screen.ViewOurAppsScreen
 import com.brandoncano.sharedcomponents.utils.OpenLink
