+++ app/src/main/java/com/brandoncano/resistancecalculator/navigation/learn/LearnPreferredValuesScreenNavEntry.kt
@@ -1,21 +1,22 @@
-package com.brandoncano.resistancecalculator.navigation
+package com.brandoncano.resistancecalculator.navigation.learn
 
 import androidx.compose.animation.slideInVertically
 import androidx.compose.animation.slideOutVertically
 import androidx.navigation.NavGraphBuilder
 import androidx.navigation.NavHostController
 import androidx.navigation.compose.composable
-import com.brandoncano.resistancecalculator.ui.screens.info.LearnSmdCodesScreen
+import com.brandoncano.resistancecalculator.navigation.Screen
+import com.brandoncano.resistancecalculator.ui.screens.info.LearnPreferredValuesScreen
 
-fun NavGraphBuilder.learnSmdCodes(
+fun NavGraphBuilder.learnPreferredValues(
     navHostController: NavHostController,
 ) {
     composable(
-        route = Screen.LearnSmdCodes.route,
+        route = Screen.LearnPreferredValues.route,
         enterTransition = { slideInVertically(initialOffsetY = { it }) },
         exitTransition = { slideOutVertically(targetOffsetY = { it }) },
     ) {
-        LearnSmdCodesScreen(
+        LearnPreferredValuesScreen(
             onNavigateBack = { navHostController.popBackStack() },
         )
     }
