+++ app/src/main/java/com/brandoncano/resistancecalculator/navigation/learn/LearnSmdCodesScreenNavEntry.kt
@@ -1,21 +1,22 @@
-package com.brandoncano.resistancecalculator.navigation
+package com.brandoncano.resistancecalculator.navigation.learn
 
 import androidx.compose.animation.slideInVertically
 import androidx.compose.animation.slideOutVertically
 import androidx.navigation.NavGraphBuilder
 import androidx.navigation.NavHostController
 import androidx.navigation.compose.composable
-import com.brandoncano.resistancecalculator.ui.screens.info.LearnColorCodesScreen
+import com.brandoncano.resistancecalculator.navigation.Screen
+import com.brandoncano.resistancecalculator.ui.screens.info.LearnSmdCodesScreen
 
-fun NavGraphBuilder.learnColorCodes(
+fun NavGraphBuilder.learnSmdCodes(
     navHostController: NavHostController,
 ) {
     composable(
-        route = Screen.LearnColorCodes.route,
+        route = Screen.LearnSmdCodes.route,
         enterTransition = { slideInVertically(initialOffsetY = { it }) },
         exitTransition = { slideOutVertically(targetOffsetY = { it }) },
     ) {
-        LearnColorCodesScreen(
+        LearnSmdCodesScreen(
             onNavigateBack = { navHostController.popBackStack() },
         )
     }
