+++ app/src/main/java/com/brandoncano/resistancecalculator/ui/screens/ctv/ColorToValueComponents.kt
@@ -21,14 +21,14 @@ import com.brandoncano.resistancecalculator.model.ctv.ResistorCtv
 import com.brandoncano.resistancecalculator.ui.theme.ResistorCalculatorTheme
 import com.brandoncano.resistancecalculator.util.ColorFinder
 import com.brandoncano.resistancecalculator.util.Sdk
-import com.brandoncano.resistancecalculator.util.bandFiveForDisplay
-import com.brandoncano.resistancecalculator.util.bandFourForDisplay
-import com.brandoncano.resistancecalculator.util.bandOneForDisplay
-import com.brandoncano.resistancecalculator.util.bandSixForDisplay
-import com.brandoncano.resistancecalculator.util.bandThreeForDisplay
-import com.brandoncano.resistancecalculator.util.bandTwoForDisplay
-import com.brandoncano.resistancecalculator.util.deriveResistorColor
-import com.brandoncano.resistancecalculator.util.formatResistance
+import com.brandoncano.resistancecalculator.util.resistor.bandFiveForDisplay
+import com.brandoncano.resistancecalculator.util.resistor.bandFourForDisplay
+import com.brandoncano.resistancecalculator.util.resistor.bandOneForDisplay
+import com.brandoncano.resistancecalculator.util.resistor.bandSixForDisplay
+import com.brandoncano.resistancecalculator.util.resistor.bandThreeForDisplay
+import com.brandoncano.resistancecalculator.util.resistor.bandTwoForDisplay
+import com.brandoncano.resistancecalculator.util.resistor.deriveResistorColor
+import com.brandoncano.resistancecalculator.util.resistor.formatResistance
 import com.brandoncano.sharedcomponents.composables.AppCard
 import com.brandoncano.sharedcomponents.composables.AppComponentPreviews
 import com.brandoncano.sharedcomponents.composables.DrawContent
