+++ app/src/main/java/com/brandoncano/resistancecalculator/util/resistor/ResistanceFormatter.kt
@@ -1,6 +1,8 @@
-package com.brandoncano.resistancecalculator.util
+package com.brandoncano.resistancecalculator.util.resistor
 
 import com.brandoncano.resistancecalculator.model.ctv.ResistorCtv
+import com.brandoncano.resistancecalculator.util.UnitsFromMultiplier
+import com.brandoncano.resistancecalculator.util.ValueFinder
 import com.brandoncano.resistancecalculator.constants.Colors as C
 import com.brandoncano.resistancecalculator.constants.Symbols as S
