+++ app/src/main/java/com/brandoncano/resistancecalculator/util/resistor/ResistorExtensions.kt
@@ -1,9 +1,10 @@
-package com.brandoncano.resistancecalculator.util
+package com.brandoncano.resistancecalculator.util.resistor
 
 import com.brandoncano.resistancecalculator.constants.Colors
 import com.brandoncano.resistancecalculator.model.ctv.ResistorCtv
 import com.brandoncano.resistancecalculator.model.smd.SmdResistor
 import com.brandoncano.resistancecalculator.model.vtc.ResistorVtc
+import com.brandoncano.resistancecalculator.util.ColorFinder
 
 // color to value
 fun ResistorCtv.bandOneForDisplay(): String {
