+++ app/src/main/java/com/brandoncano/resistancecalculator/util/resistor/ResistorFormatter.kt
@@ -1,7 +1,9 @@
-package com.brandoncano.resistancecalculator.util
+package com.brandoncano.resistancecalculator.util.resistor
 
 
 import com.brandoncano.resistancecalculator.model.vtc.ResistorVtc
+import com.brandoncano.resistancecalculator.util.ColorFinder
+import com.brandoncano.resistancecalculator.util.MultiplierFromUnits
 import java.util.Locale
 import com.brandoncano.resistancecalculator.constants.Colors as C
