+++ app/src/test/java/com/brandoncano/resistancecalculator/util/IsValidResistanceTest.kt
@@ -3,6 +3,7 @@ package com.brandoncano.resistancecalculator.util
 import android.content.Context
 import com.brandoncano.resistancecalculator.data.SharedPreferences
 import com.brandoncano.resistancecalculator.model.vtc.ResistorVtc
+import com.brandoncano.resistancecalculator.util.resistor.IsValidResistance
 import io.mockk.clearAllMocks
 import io.mockk.every
 import io.mockk.mockk
