+++ app/src/test/java/com/brandoncano/resistancecalculator/util/ResistorFormatterTest.kt
@@ -1,6 +1,7 @@
 package com.brandoncano.resistancecalculator.util
 
 import com.brandoncano.resistancecalculator.model.vtc.ResistorVtc
+import com.brandoncano.resistancecalculator.util.resistor.ResistorFormatter
 import com.brandoncano.resistancecalculator.constants.Colors as C
 import com.brandoncano.resistancecalculator.constants.Symbols as S
 import org.junit.Assert.assertEquals
